+++ core/src/main/scala/pureconfig/configurable/package.scala
@@ -18,10 +18,12 @@ import pureconfig.error.FailureReason
   *   implicit val localDateConfigConvert = makeLocalDateConfigConvert(DateTimeFormatter.ISO_TIME)
   * }}}
   *
-  * @example we cannot provide a [[ConfigConvert]] for [[java.time.LocalDate]] because traditionally there are many different
-  * [[java.time.format.DateTimeFormatter]]s to parse a [[java.time.LocalDate]] from a [[java.lang.String]]. This package
-  * provides a method that takes an input [[java.time.format.DateTimeFormatter]] and returns a [[ConfigConvert]] for
-  * [[java.time.LocalDate]] which will use that [[java.time.format.DateTimeFormatter]] to parse a [[java.time.LocalDate]].
+  * @example
+  *   we cannot provide a [[ConfigConvert]] for [[java.time.LocalDate]] because traditionally there are many different
+  *   [[java.time.format.DateTimeFormatter]] s to parse a [[java.time.LocalDate]] from a [[java.lang.String]]. This
+  *   package provides a method that takes an input [[java.time.format.DateTimeFormatter]] and returns a
+  *   [[ConfigConvert]] for [[java.time.LocalDate]] which will use that [[java.time.format.DateTimeFormatter]] to parse
+  *   a [[java.time.LocalDate]].
   */
 package object configurable {