+++ core/src/main/scala/pureconfig/error/ConfigReaderFailure.scala
@@ -28,9 +28,12 @@ trait ConfigReaderFailure {
 /** A failure occurred when converting from a `ConfigValue` to a given type. The failure contains a path to the
   * `ConfigValue` that raised the error.
   *
-  * @param reason the reason for the conversion failure
-  * @param origin the optional origin of the failure
-  * @param path the path to the `ConfigValue` that raised the error
+  * @param reason
+  *   the reason for the conversion failure
+  * @param origin
+  *   the optional origin of the failure
+  * @param path
+  *   the path to the `ConfigValue` that raised the error
   */
 case class ConvertFailure(reason: FailureReason, origin: Option[ConfigOrigin], path: String)
     extends ConfigReaderFailure {
@@ -42,26 +45,23 @@ object ConvertFailure {
 
   /** Constructs a `ConvertFailure` from a reason and a `ConfigCursor`.
     *
-    * @param reason the reason for the conversion failure
-    * @param cur the cursor where the failure ocurred
-    * @return a `ConvertFailure` for the given reason at the given cursor.
+    * @param reason
+    *   the reason for the conversion failure
+    * @param cur
+    *   the cursor where the failure ocurred
+    * @return
+    *   a `ConvertFailure` for the given reason at the given cursor.
     */
   def apply(reason: FailureReason, cur: ConfigCursor): ConvertFailure =
     ConvertFailure(reason, cur.origin, cur.path)
 }
 
-/** A failure occurred because a list of files to load was empty.
-  */
-@deprecated("`loadConfigFromFiles` won't return this failure anymore", "0.10.1")
-case object NoFilesToRead extends ConfigReaderFailure {
-  def description = "The config files to load must not be empty."
-  def origin = None
-}
-
 /** A failure occurred because an exception was thrown during the reading process.
   *
-  * @param throwable the exception thrown
-  * @param origin the optional origin of the failure
+  * @param throwable
+  *   the exception thrown
+  * @param origin
+  *   the optional origin of the failure
   */
 final case class ThrowableFailure(throwable: Throwable, origin: Option[ConfigOrigin]) extends ConfigReaderFailure {
 
@@ -98,8 +98,10 @@ trait CannotRead extends ConfigReaderFailure {
 
 /** A failure occurred due to the inability to read a requested file.
   *
-  * @param path the file system path of the file that couldn't be read
-  * @param reason an optional exception thrown when trying to read the file
+  * @param path
+  *   the file system path of the file that couldn't be read
+  * @param reason
+  *   an optional exception thrown when trying to read the file
   */
 final case class CannotReadFile(path: Path, reason: Option[Throwable]) extends CannotRead {
   val sourceType = "file"
@@ -108,8 +110,10 @@ final case class CannotReadFile(path: Path, reason: Option[Throwable]) extends C
 
 /** A failure occurred due to the inability to read a requested URL.
   *
-  * @param url the URL that couldn't be read
-  * @param reason an optional exception thrown when trying to read the URL
+  * @param url
+  *   the URL that couldn't be read
+  * @param reason
+  *   an optional exception thrown when trying to read the URL
   */
 final case class CannotReadUrl(url: URL, reason: Option[Throwable]) extends CannotRead {
   val sourceType = "URL"
@@ -118,8 +122,10 @@ final case class CannotReadUrl(url: URL, reason: Option[Throwable]) extends Cann
 
 /** A failure occurred due to the inability to read a requested resource.
   *
-  * @param resourceName the resource that couldn't be read
-  * @param reason an optional exception thrown when trying to read the resource
+  * @param resourceName
+  *   the resource that couldn't be read
+  * @param reason
+  *   an optional exception thrown when trying to read the resource
   */
 final case class CannotReadResource(resourceName: String, reason: Option[Throwable]) extends CannotRead {
   val sourceType = "resource"
@@ -128,8 +134,10 @@ final case class CannotReadResource(resourceName: String, reason: Option[Throwab
 
 /** A failure occurred due to the inability to parse the configuration.
   *
-  * @param msg the error message from the parser
-  * @param origin the optional origin of the failure
+  * @param msg
+  *   the error message from the parser
+  * @param origin
+  *   the optional origin of the failure
   */
 final case class CannotParse(msg: String, origin: Option[ConfigOrigin]) extends ConfigReaderFailure {
   def description = s"Unable to parse the configuration: $msg."