+++ example/src/main/scala/pureconfig/example/conf/Email.scala
@@ -5,8 +5,7 @@ package pureconfig.example
 
 import scala.util.{Failure, Success, Try}
 
-/** This is not a production-quality email address validator.
-  * It is provided only for illustration purposes.
+/** This is not a production-quality email address validator. It is provided only for illustration purposes.
   */
 object Email {
   private val regex = """^([a-zA-Z0-9!#$%&'.*+/=?^_`{|}~;-]+)@([a-zA-Z0-9.-]+)$""".r