+++ modules/magnolia/src/main/scala/pureconfig/module/magnolia/EnumerationConfigWriterBuilder.scala
@@ -10,7 +10,8 @@ import pureconfig.ConfigWriter
 /** A type class to build `ConfigWriter`s for sealed families of case objects where each type is encoded as a
   * `ConfigString` based on the type name.
   *
-  * @tparam A the type of objects capable of being written as an enumeration
+  * @tparam A
+  *   the type of objects capable of being written as an enumeration
   */
 private[magnolia] trait EnumerationConfigWriterBuilder[A] {
   def build(transformName: String => String): ConfigWriter[A]