+++ modules/magnolia/src/test/scala/pureconfig/module/magnolia/TupleConvertersSuite.scala
@@ -4,7 +4,7 @@ import scala.collection.JavaConverters._
 import scala.language.higherKinds
 
 import com.typesafe.config.{ConfigValueFactory, ConfigValueType}
-import org.scalacheck.ScalacheckShapeless._
+import org.scalacheck.Arbitrary
 
 import pureconfig._
 import pureconfig.error._
@@ -12,6 +12,11 @@ import pureconfig.module.magnolia.auto.reader._
 import pureconfig.module.magnolia.auto.writer._
 
 class TupleConvertersSuite extends BaseSuite {
+  case class Foo(a: Int, b: String)
+
+  implicit val arbFoo: Arbitrary[Foo] = Arbitrary {
+    Arbitrary.arbitrary[(Int, String)].map((Foo.apply _).tupled)
+  }
 
   behavior of "ConfigConvert"
 
@@ -21,7 +26,6 @@ class TupleConvertersSuite extends BaseSuite {
   checkArbitrary[(Int, (Long, String), Boolean)]
 
   // Check arbitrary Tuples with custom types
-  case class Foo(a: Int, b: String)
   checkArbitrary[(Long, Foo, Boolean, Foo)]
 
   // Check readers from objects and lists